refactor: concrete InstanceManager with factory for sub-managers
Big InstanceManager class with generic was a pain to manage and caused Zod implementation to be painful. Rather, use a factory pattern with a generic InstanceManager. Depending on State and provider in state, implement the InstanceManager with expected sub-Managers for specific provider.

This also provide  a better separation of concern between high-level InstanceManager (not knowing concrete state type or provider) and provider-specific implementation.

Also started a bit of Zod implementation. No major impact yet.