refactor: renamed provision "config" as "input" for better semantic
"config" is also used as a phase name (init => provision => configuration) so it was not adapted to have a provision "config" along with "output". Using "input" closely relate to existing "output" for provisioning, and we can use the same pattern for configuration aspects. (eg. configuration.input and configuration.output)

Inspired from Terraform and Pulumi using the same semantics.