fix: more robust logic to identify GCC version used to compile kernel
as Nvidia driver requires specific GCC version we used a fragile logic to extract GCC version used to build kernel from /proc/version, but depending on kernel output differed and task would fail.

This method should be more solid on Ubuntu OSes - may not work on other OS though. Inspired from https://unix.stackexchange.com/questions/766939/how-to-robustly-determine-which-version-of-gcc-was-used-to-build-the-kernel-from