chore: cleanup and rename internal classes
Tried to implement a DummyStateWruter for Dummy provider but stopped as it requires to have a full "backend" for State management in memory. Otherwise we would have to pick a State Writer, Initializer, etc. every time an instance is loaded or persisted. This should be a more global config. This is  not a story of only Dummy provider, but an entire need for a Dummy State backend