fix: Sunshine password special characters handling
some password with special characters were not handled properly,
such as password starting with '@' or '`' which are YAML reserved scalars.
Also the "quote" Ansible filter didn't behave as expected as it's reserved for shell
not YAML...

Now passing it all as base64 encoded string and
decoding as close as possible without passing
decoded string around.