feat: move existing data to data disk if any when a data disk is provisioned on instance after initial creation
currently instances do not have data disk by default (appart for recent Scaleway instances). As Data disk will become the default soon and to ensure retrocompatibilty on update, we need to ensure existing data are moved to new data disk when it's mounted and not "lost" as they would be hidden by mount.