feat: disable shader pre-caching by default
shader pre-caching is not needed in most cases since Cloud GPU are powerful enough to make it transparent. It was annoying to users as it may take a (very) long time to download shader cache on game launch (30+ min) and it often happens every time game is started. Multiple users rpeorted frustration, disabling it will provide a better gaming experience.